Expose the fact we've had an error
This added an 'status' field to the report table.
This allows us to expose the fact we've had an error to the UI, although
its not a very nice error atm :)
To test:
http://localhost:8000/?manager=npm&package=async&version=3.2.0
Initially this will show the 'come back later' message.
But after the job fails it will show the 'unexpected error' message.